# Conflicts:
#	src/fields/Matrix.php

Reverts 97cc9a70d7b16231bf1da1a0e2db4111ad0fc011

+++ src/fields/Matrix.php
@@ -699,11 +699,14 @@ class Matrix extends Field implements EagerLoadingFieldInterface
     /**
      * @inheritdoc
      */
-    public function afterDelete(): bool
+    public function beforeDelete(): bool
     {
-        Craft::$app->getMatrix()->deleteMatrixField($this);
+        if (!parent::beforeDelete()) {
+            return false;
+        }
 
-        return parent::afterDelete();
+        Craft::$app->getMatrix()->deleteMatrixField($this);
+        return true;
     }
 
     /**
